* ThreadsController.fs: Get POST working with Autogen api
* plw.fsproj:
* plw.userprefs:
* Thread.fs:
* FeedItem.fs:
* FeedsProvider.fs:
* ThreadsProvider.fs:
* DatabaseConnectionProvider.fs:

+++ Controllers/ThreadsController.fs
@@ -12,9 +12,9 @@ type ThreadsController () =
     inherit Controller()
 
     [<HttpGet("{id}")>]
-    member this.Get(id:int) =
+    member this.Get(id:int64) = 
         GetThread(id)
 
     [<HttpPost>]
-    member this.Post([<FromBody>]data: postData) =
+    member this.Post([<FromBody>]data: postData) = 
         CreateThread(Some data.name)
